wip(product):services

+++ backend/internal/products/model/product.go
@@ -5,11 +5,11 @@ import "github.com/google/uuid"
 type Product struct {
 	Id          uuid.UUID      `gorm:"column:id;primary_key;default:uuid_generate_v4();<-:create"`
 	Name        string         `gorm:"column:name"`
-	Price       int            `gorm:"column:price"`
+	Price       int64          `gorm:"column:price"`
 	Description string         `gorm:"column:description"`
 	CreatedAt   int64          `gorm:"column:created_at;autoCreateTime:milli;<-:create"`
 	UpdatedAt   int64          `gorm:"column:updated_at;autoCreateTime:milli;autoUpdateTime:milli"`
-	DeletedAt    *int64         `gorm:"column:delete_at"`
+	DeletedAt   *int64         `gorm:"column:delete_at"`
 	Images      []ProductImage `gorm:"foreignKey:ProductId;references:Id"`
 }
